+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added Query methods to the persistant database
+ - Added program data path for the persistant database
+
 ### Changed
  - ISPlusOriginatorShim now works with any IOriginator
  - Fixed AbstractSPlusOriginatorShim delegates to be namespace level to work with S+